Refactor convolution and max pooling instructions to use i64 for dimensions
- Updated ConvInstruction and MaxPoolInstruction structs to use Vec<i64> for dilations, kernel_shape, pads, and strides instead of Vec<usize>.
- Modified compute_pads methods in both instructions to return Vec<i64> instead of Vec<usize>.
- Adjusted the implementation of convolution and max pooling functions to handle i64 dimensions throughout.
- Refactored f32_f32_f32_f32_cpu and f32_f32_cpu functions to accept &[i64] instead of Vec<usize> for dimensions and parameters.
- Updated the calc_begin_and_end_pads function to work with i64 types for pads, kernel_shape, strides, and dilations.
- Ensured all related functions and layers (ConvLayer, MaxPoolLayer) are consistent with the new i64 type usage.

+++ src/instruction/mod.rs
@@ -65,11 +65,11 @@ pub fn conv(
     bias: Option<TensorId>,
     dst: TensorId,
     auto_pad: OnnxAutoPad,
-    dilations: Vec<usize>,
+    dilations: Vec<i64>,
     group: i64,
-    kernel_shape: Vec<usize>,
-    pads: Vec<usize>,
-    strides: Vec<usize>,
+    kernel_shape: Vec<i64>,
+    pads: Vec<i64>,
+    strides: Vec<i64>,
 ) -> Box<dyn Instruction> {
     Box::new(ConvInstruction {
         src,
@@ -121,10 +121,10 @@ pub fn maxpool(
     src: TensorId,
     dst: TensorId,
     auto_pad: OnnxAutoPad,
-    dilations: Vec<usize>,
-    kernel_shape: Vec<usize>,
-    pads: Vec<usize>,
-    strides: Vec<usize>,
+    dilations: Vec<i64>,
+    kernel_shape: Vec<i64>,
+    pads: Vec<i64>,
+    strides: Vec<i64>,
     ceil_mode: bool,
 ) -> Box<dyn Instruction> {
     Box::new(MaxPoolInstruction {
